+++ host/arm/media_player_protocol.h
@@ -15,9 +15,9 @@
 
 #define MEDIA_PLAYER_CAPABILITIES \
     "protocol=1 sources=file reserved_sources=dvd " \
-    "containers=m2v,mpeg-ps,mp3 video=h262 " \
+    "containers=m2v,mpeg-ps,mp3,wav video=h262 " \
     "audio=mp2-s16le-44100,mp2-s16le-48000," \
-    "mp3-s16le-44100,mp3-s16le-48000 " \
+    "mp3-s16le-44100,mp3-s16le-48000,wav-s16le-stereo-44100,wav-s16le-stereo-48000 " \
     "transport=inband-pcm-v1"
 
 #endif